Move CommonApp to app-common and rename to BaseApplication
The dependencies are now declared within AppCommonModule instead of the BaseApplication
