+++ app-common/build.gradle.kts
@@ -8,13 +8,10 @@ android {
 
 dependencies {
     api(projects.legacy.common)
-
     api(projects.legacy.ui.legacy)
 
     api(projects.feature.account.core)
-
     api(projects.feature.launcher)
-
     api(projects.feature.navigation.drawer.api)
 
     implementation(projects.legacy.core)
@@ -27,6 +24,9 @@ dependencies {
 
     implementation(projects.feature.account.setup)
     implementation(projects.feature.migration.provider)
+    implementation(projects.feature.widget.messageList)
 
     implementation(projects.mail.protocols.imap)
+
+    implementation(libs.androidx.work.runtime)
 }
