+++ app-thunderbird/src/main/kotlin/net/thunderbird/android/ThunderbirdApp.kt
@@ -1,12 +1,12 @@
 package net.thunderbird.android
 
 import app.k9mail.feature.telemetry.api.TelemetryManager
-import com.fsck.k9.CommonApp
 import com.fsck.k9.K9
+import net.thunderbird.app.common.BaseApplication
 import org.koin.android.ext.android.inject
 import org.koin.core.module.Module
 
-class ThunderbirdApp : CommonApp() {
+class ThunderbirdApp : BaseApplication() {
     private val telemetryManager: TelemetryManager by inject()
 
     override fun provideAppModule(): Module = appModule
