+++ app-thunderbird/src/test/kotlin/net/thunderbird/android/DependencyInjectionTest.kt
@@ -24,10 +24,8 @@ import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
-import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(application = TestApp::class)
 class DependencyInjectionTest : AutoCloseKoinTest() {
     private val lifecycleOwner = mock<LifecycleOwner> {
         on { lifecycle } doReturn mock()
